+++ programs/bpf/rust/clock/Cargo.toml
@@ -2,9 +2,9 @@
 # Note: This crate must be built using build.sh
 
 [package]
-name = "solana-bpf-rust-tick-height"
+name = "solana-bpf-rust-clock"
 version = "0.19.0-pre0"
-description = "Solana BPF noop program written in Rust"
+description = "Solana BPF clock sysvar test"
 authors = ["Solana Maintainers <maintainers@solana.com>"]
 repository = "https://github.com/solana-labs/solana"
 license = "Apache-2.0"
@@ -12,7 +12,6 @@ homepage = "https://solana.com/"
 edition = "2018"
 
 [dependencies]
-byteorder = { version = "1", default-features = false }
 solana-sdk = { path = "../../../../sdk/", version = "0.19.0-pre0", default-features = false }
 
 [features]
@@ -24,4 +23,4 @@ members = []
 
 [lib]
 crate-type = ["cdylib"]
-name = "solana_bpf_rust_tick_height"
+name = "solana_bpf_rust_clock"
